Refactor & break Post file into types and helpers
