+++ packages/web/components/Dashboard/Post/helpers.ts
@@ -1,6 +1,9 @@
 import { PostClapFragmentFragment as PostClapType } from '@/generated/graphql'
 import { useTranslation } from '@/config/i18n'
 
+// Elements whose boundaries a comment can cross
+const elementWhiteList = new Set(['SPAN', 'EM', 'U', 'STRONG'])
+
 /**
  * Get an element’s position relative to the document
  * @see https://stackoverflow.com/a/26230989/3610495
@@ -67,4 +70,114 @@ export const getUsersClappedText = (
       numOthers: numOthersClapped
     })
   }
+}
+
+// Construct highlighted text/selection & replace original selection with highlighted construction
+export function highlightRange(range: Range, threadId: number): string {
+  const selectedText = range.extractContents()
+  const commentedTextSpan = document.createElement('span')
+  commentedTextSpan.classList.add('thread-highlight')
+  commentedTextSpan.dataset.tid = `${threadId}`
+  commentedTextSpan.appendChild(selectedText)
+  range.insertNode(commentedTextSpan)
+
+  return commentedTextSpan.innerHTML
+}
+
+export function buildPreOrderList(rootEl: HTMLElement): (HTMLElement | Node)[] {
+  const preOrderList: (HTMLElement | Node)[] = []
+  const recur = (el: HTMLElement | Node) => {
+    preOrderList.push(el)
+    el.childNodes.forEach(recur)
+  }
+  recur(rootEl)
+  return preOrderList
+}
+
+export function buildPostOrderList(el: HTMLElement): (HTMLElement | Node)[] {
+  const postOrderList: (HTMLElement | Node)[] = []
+  const recur = (el: HTMLElement | Node) => {
+    el.childNodes.forEach(recur)
+    postOrderList.push(el)
+  }
+  recur(el)
+  return postOrderList
+}
+
+// Returns boolean to indicate whether selection is valid for a comment
+export function isSelectionCommentable(selection: Selection, parentElement: HTMLElement) {
+  if (
+    !selection.anchorNode ||
+    !selection.focusNode ||
+    selection.isCollapsed ||
+    !selection.toString().trim().length ||
+    !isChildOf(selection.anchorNode, parentElement) ||
+    !isChildOf(selection.focusNode, parentElement)
+  ) {
+    return false
+  }
+
+  const nodeList = buildPostOrderList(parentElement)
+  // Index of the element that contains the beginning of the selection
+  let startIdx = nodeList.indexOf(selection.anchorNode)
+  // Index of the element that contains the end of the selection
+  let endIdx = nodeList.indexOf(selection.focusNode)
+
+  // Make sure it's not a backwards selection
+  // If so, reverse it.
+  if (startIdx > endIdx) {
+    ;[startIdx, endIdx] = [endIdx, startIdx]
+  }
+
+  // The nodes that the selection crosses boundaries of
+  const selectedNodes = nodeList.slice(startIdx, endIdx + 1)
+
+  // Check none of the nodes crossed by the selection
+  // are not Text nodes or outisde of the whitelist
+  for (const node of selectedNodes) {
+    if (node.constructor === Text) {
+      continue
+    } else if (elementWhiteList.has((node as HTMLElement).tagName)) {
+      continue
+    } else {
+      // node not in our "whitelist"
+      return false
+    }
+  }
+  return true
+}
+
+export function buildPreOrderListAndOffsets(selectableTextArea: HTMLElement) {
+  // a list of every element in the selectableTextArea
+  const preOrderList = buildPreOrderList(selectableTextArea)
+  // a list of integers which are the offsets of the preOrderList elements
+  // this method of creating the Array is more efficient than using Array.push
+  const offsets = new Array(preOrderList.length)
+
+  // Offset of the sum of the length of all the Text Nodes that appear
+  // before the current element in the document
+  let currentOffset = 0
+  for (let i = 0; i < preOrderList.length; i++) {
+    const node = preOrderList[i]
+    offsets[i] = currentOffset
+
+    if (node.constructor === Text) {
+      currentOffset += (node as Text).length
+    }
+  }
+
+  return [preOrderList, offsets]
+}
+
+export function isChildOf(el: Node, target: HTMLElement) {
+  let current = el
+  while (current.parentElement) {
+    if (current.parentElement === target) {
+      return true
+    }
+
+    current = current.parentElement
+  }
+
+  return false
 }
